models: group by group uid
This is a first step required to properly render channels or
multi-person chats. Before this commit, the channels were just not
rendered at all and the individual messages in the channels were shown
as direct messages.

The window title still needs to be fixed, and currently the usernames
within a channel window is set to just the channel name. The logging
code will also need to set a group title/name, and not just a group_uid,
but we will get to that as well.