tp: make channel sending work
The previous commit didn't work for rooms, but now it does.
During testing I also exposed an edge case where the remote_uid was
"Wizzup" but the open channel was to "wizzup", we'll have to deal with
this properly somehow, for now I just used QString's .toLower() to work
around this problem, but that might not be the case for all protocols.