Support opening chats/windows from Address book
This makes it possible to open a chat with any contact using the address
book. The side effect of this is that it will also open a channel window
if joinChannel() is called, but I think that should be a good thing, not
a bad thing.

Also refactor some of the TP code a bit to remove code duplication. The
pointer handling should be cleaned up a bit to just use the
TextChannelPtr variants rather than TextChannel*.