notify: add sounds, vibration, grouping
Also discern between SMS and chat

Also remove not showing notifications when the program is backgrounded,
we do not to change this a bit later to not show a notification if the
*active* window is the one we're notifying for.  Also remove the rate
limit, since Fremantle did not have it.

The grouping needs more work, events aren't persistent yet (this
requires us to add some more notification handling). libnotify-qt seems
to be a little buggy and there is a bunch of stuff missing, so we'll
have to fix it up further.