rtcom: fix 'message_read' for preceding messages in current buffer relative to most recent message
Removed QML functionality to detect if messages are "in view". Instead, the ChatWindow constructor now sets messages as read, as opening a chat window implies the messages are about to be read.

Moved rtcom logic related to ChatMessage to ChatModel, as having it in rtcom.h introduces unnecesary circular dependencies. rtcom.h now contains only rtcom functionality.

Same goes for utils, since utils.h is used virtually everywhere the combination of Utils and Notification created a circular dependency. Moved the creation of notifications to the static Notification::issue().